requirements.txt
fastapi
uvicorn
weaviate-client==4.8.1
numpy
httpx
pydantic